Refactor: config 관련 파일을 config 폴더 하위로 이동
