+++ src/api/auth/auth.module.ts
@@ -3,8 +3,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 
+import { ServerConfig } from '@src/config/server/server.type';
+import { SERVER_CONFIG_TOKEN } from '@src/config/server/server.constant';
 import { User } from '@src/entity/user.entity';
-import { ServerConfig } from '@src/shared/interface/config.interface';
 
 import { AuthService } from './auth.service';
 import { JwtStrategy } from './strategy/jwt.strategy';
@@ -16,7 +17,7 @@ import { AuthController } from './auth.controller';
     JwtModule.registerAsync({
       inject: [ConfigService],
       useFactory: (configService: ConfigService) => {
-        const { jwtSecret, jwtExpiresIn } = configService.get<ServerConfig>('server');
+        const { jwtSecret, jwtExpiresIn } = configService.get<ServerConfig>(SERVER_CONFIG_TOKEN);
 
         return {
           secret: jwtSecret,
