+++ src/config/db/db.constant.ts
@@ -1,0 +1,1 @@
+export const DB_CONFIG_TOKEN = 'db';
