+++ src/config/db/db.type.ts
@@ -1,0 +1,7 @@
+export interface DbConfig {
+  host: string;
+  port: number;
+  username: string;
+  password: string;
+  database: string;
+}
