+++ src/config/server/server.constant.ts
@@ -1,0 +1,1 @@
+export const SERVER_CONFIG_TOKEN = 'server';
