+++ src/main.ts
@@ -1,13 +1,15 @@
 import { NestFactory } from '@nestjs/core';
 import { ConfigService } from '@nestjs/config';
 import * as cookieParser from 'cookie-parser';
+
+import { ServerConfig } from './config/server/server.type';
+import { SERVER_CONFIG_TOKEN } from './config/server/server.constant';
 import { AppModule } from './app.module';
-import { ServerConfig } from './shared/interface/config.interface';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   const configService = app.get<ConfigService>(ConfigService);
-  const { port, domain } = configService.get<ServerConfig>('server');
+  const { port, domain } = configService.get<ServerConfig>(SERVER_CONFIG_TOKEN);
 
   app.enableCors({
     origin: domain, // 허용할 도메인
