--- src/shared/enum/node-env.enum.ts
@@ -1,5 +1,0 @@
-export enum NodeEnv {
-  DEV = 'dev',
-  PROD = 'prod',
-  TEST = 'test',
-}
